+++ components/navigation.js
@@ -10,18 +10,19 @@ export default function Navigation () {
         <div className={styles.navigationMenu}>
             <div className="logo">
                 <Link href="/">
-                    <Image src="/assets/alluvium-logo.svg" width={150} height={29.06} alt="logo"/>
+                    <img src="/assets/alluvium-logo.svg" alt="logo"/>
+                    {/* <Image src="/assets/alluvium-logo.svg" width={150} height={29.06} alt="logo"/> */}
                 </Link>
             </div>
             <nav className="navigation">
                 <ul className="desktop-menu">
-                    <li><Link href="/projects">Our Projects</Link></li>
+                    <li><Link href="/products">Our Projects</Link></li>
                     <li><Link href="/reel">Project Reel</Link></li>
                     <li><Link href="/people">Alluvians</Link></li>
                     <li><Link href="/support">Support</Link></li>
                 </ul>
                 <ul className={styles.mobileMenu}>
-                    <li><Link href="/projects">Our Projects</Link></li>
+                    <li><Link href="/products">Our Projects</Link></li>
                     <li><Link href="/reel">Project Reel</Link></li>
                     <li><Link href="/people">Alluvians</Link></li>
                     <li><Link href="/support">Support</Link></li>
