+++ data.js
@@ -1,0 +1,230 @@
+export const productData = 
+[
+    {
+      id:1,
+      title:"Cloud Counter",
+      subtitle:"Employ our fast and efficient cloud counter to keep count of your resources",
+      imgLink:"/assets/Secure.svg",
+      projectName:"cloud_counter",
+      details:{
+        productName:"Cloud Counter",
+        productTitle:"Resource count? No worries.",
+        productSubtitle: "Get The Most out of Jira Software Cloud with our Quick and Easy Resource Counter.",
+        productLead:"Taiwo Ojo",
+        client:"alluvium",
+        date: "September 2020",
+        productImg: '/assets/test.png',
+        productInfo: "Sollicitudin risus aenean et, tellus tellus hendrerit senectus in. Quis senectus quam sed nunc id vulputate purus. Volutpat sit pretium a vitae. Eget sapien leo purus non libero, purus nunc, sit sed. In lobortis morbi at cras ultrices tristique. Aliquam in proin sed ultricies dignissim orci aliquet. Sagittis, massa nam dui quis est praesent. Eu quisque sed tincidunt nisl. At nulla blandit urna aenean. Diam auctor libero egestas fermentum metus condimentum nisi aenean urna. Ultrices iaculis leo id tempus magna orci. Eleifend risus, integer mauris consectetur at purus facilisis proin facilisi. Convallis egestas porta ullamcorper habitant viverra feugiat. Pellentesque orci consequat porttitor aenean sit nam velit viverra rhoncus. Velit, enim et massa sed diam neque amet aenean interdum. Et, in amet tempor, dui, tempus blandit sit egestas. Eu sollicitudin quisque neque, mi duis tellus ornare fermentum. Tempor nibh sed proin amet enim ut egestas gravida pretium. Et vehicula semper quam nulla. Orci orci eget tellus id. Tincidunt imperdiet dui etiam tellus. Aenean adipiscing commodo, augue purus in donec tellus augue. Sit tellus vitae rhoncus sed at urna, nibh vitae habitant. Elit, rhoncus, fringilla aenean mauris nunc sapien nibh nunc. Mauris pulvinar est venenatis diam pharetra venenatis suspendisse donec. Ac risus etiam pretium tincidunt. Mauris sapien fames senectus elementum phasellus sed nec at. Faucibus diam est eu et in eget potenti risus adipiscing. Eu enim auctor non dignissim. Velit ac at morbi vitae gravida adipiscing elit odio cras. Ultrices tristique ullamcorper lorem egestas etiam a. Tristique ullamcorper",
+        moreImgs: [ "/assets/test.png","/assets/test.png" ],
+        keyPoints: [
+          "Quis nisl elementum nunc in. Egestas urna consectetur risus neque, euismod aliquam dictum. Mi at scelerisque tristique",
+          "Matoque vulputate pellentesque felis. Dignissim aliquet amet, sit augue diam. Massa hac lorem iaculis senectus diam feugiat tristique.",
+          "Tugue purus in donec tellus augue. Sit tellus vitae rho"
+        ],
+      },
+    },
+    {
+      id:2,
+      title:"CJR: Compare Jira Resources",
+      subtitle:"Compare Jira Resources is an efficient way to check that you have fully migrated",
+      imgLink:"/assets/Simple.svg",
+      projectName:"cjr",
+      details:{
+        productName:"CJR",
+        productTitle:"Compare Jira Resources",
+        productSubtitle: "Validate success of your migration, Compare Source & For Accuracy of Jira.",
+        productLead:"Taiwo Ojo",
+        client:"alluvium",
+        date: "November 2020",
+        productImg: '/assets/test.png',
+        productInfo: "Sollicitudin risus aenean et, tellus tellus hendrerit senectus in. Quis senectus quam sed nunc id vulputate purus. Volutpat sit pretium a vitae. Eget sapien leo purus non libero, purus nunc, sit sed. In lobortis morbi at cras ultrices tristique. Aliquam in proin sed ultricies dignissim orci aliquet. Sagittis, massa nam dui quis est praesent. Eu quisque sed tincidunt nisl. At nulla blandit urna aenean. Diam auctor libero egestas fermentum metus condimentum nisi aenean urna. Ultrices iaculis leo id tempus magna orci. Eleifend risus, integer mauris consectetur at purus facilisis proin facilisi. Convallis egestas porta ullamcorper habitant viverra feugiat. Pellentesque orci consequat porttitor aenean sit nam velit viverra rhoncus. Velit, enim et massa sed diam neque amet aenean interdum. Et, in amet tempor, dui, tempus blandit sit egestas. Eu sollicitudin quisque neque, mi duis tellus ornare fermentum. Tempor nibh sed proin amet enim ut egestas gravida pretium. Et vehicula semper quam nulla. Orci orci eget tellus id. Tincidunt imperdiet dui etiam tellus. Aenean adipiscing commodo, augue purus in donec tellus augue. Sit tellus vitae rhoncus sed at urna, nibh vitae habitant. Elit, rhoncus, fringilla aenean mauris nunc sapien nibh nunc. Mauris pulvinar est venenatis diam pharetra venenatis suspendisse donec. Ac risus etiam pretium tincidunt. Mauris sapien fames senectus elementum phasellus sed nec at. Faucibus diam est eu et in eget potenti risus adipiscing. Eu enim auctor non dignissim. Velit ac at morbi vitae gravida adipiscing elit odio cras. Ultrices tristique ullamcorper lorem egestas etiam a. Tristique ullamcorper",
+        moreImgs: [ "/assets/test.png","/assets/test.png" ],
+        keyPoints: [
+          "Quis nisl elementum nunc in. Egestas urna consectetur risus neque, euismod aliquam dictum. Mi at scelerisque tristique",
+          "Matoque vulputate pellentesque felis. Dignissim aliquet amet, sit augue diam. Massa hac lorem iaculis senectus diam feugiat tristique.",
+          "Tugue purus in donec tellus augue. Sit tellus vitae rho"
+        ],
+      },
+    },
+]
+
+
+export const featuresData = 
+[
+    {
+        title:"Secure",
+        subtitle:"Data Security is so important to us that Everyone in our company is GDPR certified, even the janitor.",
+        iconLink:"/assets/Secure.svg"
+    },
+    {
+        title:"Simple-UX",
+        subtitle:"We do our best to make sure the customer do not notice changes in their user experience, post migration",
+        iconLink:"/assets/Simple.svg"
+    },
+    {
+        title:"Scalable",
+        subtitle:"Build, rinse, repeat",
+        iconLink:"/assets/Scalable.svg"
+    },
+]
+  
+export const previewData = 
+[
+    {
+      title:"Cloud Counter",
+      subtitle:"Employ our fast and efficient cloud counter to keep count of your resources ",
+      imgLink:"/assets/Secure.svg",
+      projectName:"cloud_counter"
+    },
+    {
+      title:"CJR: Compare Jira Resources",
+      subtitle:"Compare Jira Resources is an efficient way to check that you have fully migrated",
+      imgLink:"/assets/Simple.svg",
+      projectName:"cjr"
+    },
+]
+
+export const galleryLinks = 
+[
+    {
+      imgName:"/assets/featuresImg.png",
+      imgAlt:"reel",
+      width:200,
+      height:200
+    },
+    {
+      imgName:"/assets/featuresImg.png",
+      imgAlt:"reel",
+      width:200,
+      height:200
+    },
+]
+
+export const alluviansData = 
+[
+    {
+      firstName:"taiwo",
+      lastName:"ojo",
+      role:"Founder/Team lead",
+      email:"taiwo@alluvium.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"ibukun",
+      lastName:"akilo",
+      role:"Founder/Team Lead",
+      email:"ibukunakilo@gmail.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"enyioma",
+      lastName:"osondu",
+      role:"Backend engineer",
+      email:"enyioma.osondu@geniesys.co.uk",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"stanley",
+      lastName:"onwuka",
+      role:"Frontend engineer",
+      email:"stanley@geniesys.co.uk",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"ayodamola",
+      lastName:"alonge",
+      role:"UI/UX Designer",
+      email:"pelumi@geniesys.co.uk",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"damilola",
+      lastName:"daramola",
+      role:"Employee",
+      email:"daramola@geniesys.co.uk",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"adeola",
+      lastName:"lawal",
+      role:"Employee",
+      email:"adeola@geniesys.co.uk",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"isaac",
+      lastName:"temidayo",
+      role:"Employee",
+      email:"isaac@geniesys.co.uk",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"precious",
+      lastName:"charles",
+      role:"Employee",
+      email:"charlespreciousa@gmail.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"francis",
+      lastName:"okeke",
+      role:"Employee",
+      email:"okeke98@gmail.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"idiongette",
+      lastName:"enimini",
+      role:"Employee",
+      email:"idiongette@alluvium.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"yusuf",
+      lastName:"braimoh",
+      role:"Employee",
+      email:"yusuf@alluvium.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"georgette",
+      lastName:"idrissou",
+      role:"Employee",
+      email:"georgette@alluvium.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"anjola",
+      lastName:"oyelakin",
+      role:"Employee",
+      email:"anjolly@alluvium.com",
+      imgUrl:"/assets/test.png"
+    },
+    {
+      firstName:"adegboyega",
+      lastName:"adesanya",
+      role:"Employee",
+      email:"adeadesanya2015@gmail.com",
+      imgUrl:"/assets/test.png"
+    },
+]
+
+export const faqData = 
+[
+    {
+      question:"Do you work for European brands only?",
+      answer:"No! Our services are extended to other parts of the world. We have good success running through Africa, parts of America and other continents."
+    },
+    {
+      question:"How long does it take to complete projects?",
+      answer:"No! Our services are extended to other parts of the world. We have good success running through Africa, parts of America and other continents."
+    },
+    {
+      question:"How much does it cost to develop a software?",
+      answer:"No! Our services are extended to other parts of the world. We have good success running through Africa, parts of America and other continents."
+    },
+    {
+      question:"Can i get my job done in a week?",
+      answer:"No! Our services are extended to other parts of the world. We have good success running through Africa, parts of America and other continents."
+    },
+]
