+++ pages/api/faqs/index.js
@@ -1,0 +1,5 @@
+import { faqData } from '../../../data'
+
+export default function handler(req,res){
+    res.status(200).json(faqData);
+}
