+++ pages/api/features/index.js
@@ -1,0 +1,5 @@
+import { featuresData } from '../../../data'
+
+export default function handler(req,res){
+    res.status(200).json(featuresData);
+}
