+++ pages/api/gallery/index.js
@@ -1,0 +1,5 @@
+import { galleryLinks } from '../../../data';
+
+export default function handler(req,res){
+    res.status(200).json(galleryLinks);
+}
