+++ pages/api/people/index.js
@@ -1,0 +1,5 @@
+import { alluviansData } from '../../../data';
+
+export default function handler(req,res){
+    res.status(200).json(alluviansData);
+}
