+++ pages/api/products/index.js
@@ -1,0 +1,5 @@
+import { productData } from '../../../data';
+
+export default function handler(req,res){
+    res.status(200).json(productData);
+}
