--- pages/contact/index.js
@@ -1,15 +1,0 @@
-import Head from 'next/head'
-
-export default function Contact() {
-  return (
-    <div>
-      <Head>
-        <title>Alluvium</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>
-            Contact
-      </main>
-    </div>
-  )
-}
