--- pages/enquiries/index.js
@@ -1,15 +1,0 @@
-import Head from 'next/head'
-
-export default function Enquiries() {
-  return (
-    <div>
-      <Head>
-        <title>Alluvium</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>
-            Enquiries
-      </main>
-    </div>
-  )
-}
